Move actor-related traits and lifecycle management to separate module for better organization
